+++ packages/ramps-controller/src/RampsController.ts
@@ -642,7 +642,6 @@ type AllowedActions =
   | UserStorageController.UserStorageControllerGetStateAction
   | UserStorageController.UserStorageControllerPerformGetStorageAllFeatureEntriesAction
   | UserStorageController.UserStorageControllerPerformBatchSetStorageAction
-  | UserStorageController.UserStorageControllerListEntropySourcesAction
   | AuthenticationController.AuthenticationControllerIsSignedInAction;
 
 /**